+++ test/test-scripts/cmd-runDebug.js
@@ -1,0 +1,8 @@
+let { cmd } = require("../../build");
+
+let program = `
+  console.log('this is from stdout');
+  console.error('this is from stderr');
+`;
+
+cmd.text(program).pipe("node", "-").runDebug();