+++ CHANGELOG.md
@@ -13,7 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add Cmd.runSilent() to run commands without stdin, stdout or stderr.
+- Add `Cmd.runSilent()` to run commands without stdin, stdout or stderr.
+- Add `Cmd.getAll()` to capture stdout and stderr as strings.
+- Add `Cmd.runDebug()` to run commands and forward stdout + stderr.
 
 ## [0.0.0] - 2022-09-08